+++ internal/app/service/earmark_test.go
@@ -200,9 +200,9 @@ func TestService_GetEarmarksPaginated(t *testing.T) {
 		earmarks, pagination, err := svc.GetEarmarksPaginated(ctx, userID, limit, offset, archived)
 		assert.NilError(t, err)
 		assert.Equal(t, len(earmarks), currentCount)
-		assert.Equal(t, pagination.Limit, uint32(limit))
-		assert.Equal(t, pagination.Offset, uint32(offset))
-		assert.Equal(t, pagination.Count, uint32(currentCount))
+		assert.Equal(t, pagination.Limit, limit)
+		assert.Equal(t, pagination.Offset, offset)
+		assert.Equal(t, pagination.Count, currentCount)
 		// we make sure that all expectations were met
 		assert.Assert(t, mock.ExpectationsWereMet(),
 			"there were unfulfilled expectations")
@@ -255,9 +255,9 @@ func TestService_GetEarmarksPaginated(t *testing.T) {
 		earmarks, pagination, err := svc.GetEarmarksPaginated(ctx, userID, limit, offset, archived)
 		assert.NilError(t, err)
 		assert.Equal(t, len(earmarks), archivedCount)
-		assert.Equal(t, pagination.Limit, uint32(limit))
-		assert.Equal(t, pagination.Offset, uint32(offset))
-		assert.Equal(t, pagination.Count, uint32(archivedCount))
+		assert.Equal(t, pagination.Limit, limit)
+		assert.Equal(t, pagination.Offset, offset)
+		assert.Equal(t, pagination.Count, archivedCount)
 		// we make sure that all expectations were met
 		assert.Assert(t, mock.ExpectationsWereMet(),
 			"there were unfulfilled expectations")
@@ -286,9 +286,9 @@ func TestService_GetEarmarksPaginated(t *testing.T) {
 		earmarks, pagination, err := svc.GetEarmarksPaginated(ctx, userID, limit, offset, archived)
 		assert.NilError(t, err)
 		assert.Equal(t, len(earmarks), currentCount)
-		assert.Equal(t, pagination.Limit, uint32(limit))
-		assert.Equal(t, pagination.Offset, uint32(offset))
-		assert.Equal(t, pagination.Count, uint32(currentCount))
+		assert.Equal(t, pagination.Limit, limit)
+		assert.Equal(t, pagination.Offset, offset)
+		assert.Equal(t, pagination.Count, currentCount)
 		// we make sure that all expectations were met
 		assert.Assert(t, mock.ExpectationsWereMet(),
 			"there were unfulfilled expectations")
