+++ internal/app/service/favorite.go
@@ -101,9 +101,9 @@ func (s *Service) GetFavoriteEventsPaginated(
 		events = favs
 	}
 	pagination := &Pagination{
-		Limit:  uint32(limit),
-		Offset: uint32(offset),
-		Count:  uint32(count),
+		Limit:  limit,
+		Offset: offset,
+		Count:  count,
 	}
 	return events, pagination, nil
 }
